+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/edit/media/EditInlineMessageMedia.kt
@@ -15,14 +15,3 @@ suspend fun TelegramBot.editMessageMedia(
     media: TelegramMedia,
     replyMarkup: InlineKeyboardMarkup? = null
 ) = execute(EditInlineMessageMedia(inlineMessageId, media, replyMarkup))
-
-/**
- * @param replyMarkup Some [InlineKeyboardMarkup]. See [dev.inmo.tgbotapi.extensions.utils.types.buttons.inlineKeyboard]
- * as a builder for that
- */
-@Deprecated("Renamed", ReplaceWith("this.editMessageMedia(inlineMessageId, media, replyMarkup)", "dev.inmo.tgbotapi.extensions.api.edit.media.editMessageMedia"))
-suspend fun TelegramBot.editMessageCaption(
-    inlineMessageId: InlineMessageIdentifier,
-    media: TelegramMedia,
-    replyMarkup: InlineKeyboardMarkup? = null
-) = execute(EditInlineMessageMedia(inlineMessageId, media, replyMarkup))