+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/send/abstracts/ThumbedSendMessageRequest.kt
@@ -2,8 +2,4 @@ package dev.inmo.tgbotapi.requests.send.abstracts
 
 interface ThumbedSendMessageRequest<T: Any>: SendMessageRequest<T> {
     val thumbnail: String?
-
-    @Deprecated("Renamed in telegram bot api", ReplaceWith("thumbnail"))
-    val thumb: String?
-        get() = thumbnail
 }