+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/RequestError.kt
@@ -1,6 +1,6 @@
 package dev.inmo.tgbotapi.types
 
-import com.soywiz.klock.DateTime
+import korlibs.time.DateTime
 
 sealed class RequestError
 
@@ -10,7 +10,7 @@ data class RetryAfterError(
 ) : RequestError() {
     val canContinue = (seconds * 1000L) + startCountingMillis
     val leftToRetry: Long
-        get() = canContinue - DateTime.nowUnixLong()
+        get() = canContinue - DateTime.nowUnixMillisLong()
 }
 
 data class MigrateChatId(