+++ tgbotapi.utils/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/utils/shortcuts/PollCloseShortcuts.kt
@@ -1,7 +1,7 @@
 package dev.inmo.tgbotapi.extensions.utils.shortcuts
 
-import com.soywiz.klock.DateTime
-import com.soywiz.klock.TimeSpan
+import korlibs.time.DateTime
+import korlibs.time.TimeSpan
 import dev.inmo.tgbotapi.types.LongSeconds
 import dev.inmo.tgbotapi.types.Seconds
 import dev.inmo.tgbotapi.types.polls.ApproximateScheduledCloseInfo